// src/app/core/enums/font-weight.enum.ts
export enum FontWeight {
  LIGHTER = 'lighter',
  NORMAL = 'normal',
  BOLD = 'bold',
  BOLDER = 'bolder',
  ONE_HUNDRED = '100',
  TWO_HUNDRED = '200',
  THREE_HUNDRED = '300',
  FOUR_HUNDRED = '400',
  FIVE_HUNDRED = '500',
  SIX_HUNDRED = '600',
  SEVEN_HUNDRED = '700',
  EIGHT_HUNDRED = '800',
  NINE_HUNDRED = '900'
}